requirements.txt
accelerate==0.23.0
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
dnspython==2.4.2
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.36
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.17.2
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.7.3
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numpy==1.24.4
open-clip-torch==2.20.0
packaging==23.1
paella==0.0.1.dev1
pandas==2.0.3
pathtools==0.1.2
Pillow==9.3.0
protobuf==3.20.3
psutil==5.9.5
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
regex==2023.8.8
rejson==0.5.6
requests==2.31.0
safetensors==0.3.3
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.11.1
timm==0.9.7
tokenizers==0.13.3
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tqdm==4.66.1
transformers==4.33.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
uuid==1.30
uvicorn==0.23.2
wandb==0.15.10
warmup-scheduler==0.3
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
